+++ CHANGELOG.md
@@ -12,6 +12,7 @@ unreleased
 
 ### Improvements
 
+- Delete a feature alarting inconsistant encoding declaration in document on saving.
 - Avoid switching to inactive tabbed window only to show a syntax highlighting indicator.